move pizza router tests
