--- components/.babelrc
@@ -1,5 +1,0 @@
-{
-  "plugins": [
-    "transform-object-rest-spread"
-  ]
-}
