+++ cypress/component/router-example/PizzaShop/Home.vue
