+++ cypress/component/router-example/PizzaShop/index.vue
