+++ cypress/component/router-example/PizzaShop/router.js
