+++ cypress/component/router-example/PizzaShop/toppings.js
