+++ src/index.js
@@ -7,49 +7,49 @@ const { stripIndent } = require('common-tags')
 const defaultOptions = ['html', 'vue', 'base', 'mountId', 'extensions']
 
 // default mount point element ID for root Vue instance
-const defaultMountId = 'app'
+// const defaultMountId = 'app'
 
-const parentDocument = window.parent.document
-const projectName = Cypress.config('projectName')
-const appIframeId = `Your App: '${projectName}'`
-const appIframe = parentDocument.getElementById(appIframeId)
+// const parentDocument = window.parent.document
+// const projectName = Cypress.config('projectName')
+// const appIframeId = `Your App: '${projectName}'`
+// const appIframe = parentDocument.getElementById(appIframeId)
 
-function checkMountModeEnabled() {
+function checkMountModeEnabled () {
   // @ts-ignore
   if (Cypress.spec.specType !== 'component') {
     throw new Error(
-      `In order to use mount or unmount functions please place the spec in component folder`,
+      `In order to use mount or unmount functions please place the spec in component folder`
     )
   }
 }
 // having weak reference to styles prevents garbage collection
 // and "losing" styles when the next test starts
-const stylesCache = new Map()
+// const stylesCache = new Map()
 
-const copyStyles = component => {
-  let styles
-  if (stylesCache.has(component)) {
-    styles = stylesCache.get(component)
-  } else {
-    styles = document.querySelectorAll('head style')
-    if (styles.length) {
-      console.log('injected %d styles', styles.length)
-      stylesCache.set(component, styles)
-    } else {
-      console.log('No styles injected for this component')
-      styles = null
-    }
-  }
+// const copyStyles = component => {
+//   let styles
+//   if (stylesCache.has(component)) {
+//     styles = stylesCache.get(component)
+//   } else {
+//     styles = document.querySelectorAll('head style')
+//     if (styles.length) {
+//       console.log('injected %d styles', styles.length)
+//       stylesCache.set(component, styles)
+//     } else {
+//       console.log('No styles injected for this component')
+//       styles = null
+//     }
+//   }
 
-  if (!styles) {
-    return
-  }
+//   if (!styles) {
+//     return
+//   }
 
-  const head = appIframe.contentDocument.querySelector('head')
-  styles.forEach(style => {
-    head.appendChild(style)
-  })
-}
+//   const head = appIframe.contentDocument.querySelector('head')
+//   styles.forEach(style => {
+//     head.appendChild(style)
+//   })
+// }
 
 const deleteConstructor = comp => delete comp._Ctor
 
@@ -59,22 +59,6 @@ const deleteCachedConstructors = component => {
   }
   Cypress._.values(component.components).forEach(deleteConstructor)
 }
-
-// const getPageHTML = options => {
-//   return (
-//     options.html ||
-//     stripIndent`
-//     <html>
-//       <head>
-//         ${options.base ? `<base href="${options.base}" />` : ''}
-//       </head>
-//       <body>
-//         <div id="${options.mountId || defaultMountId}"></div>
-//       </body>
-//     </html>
-//   `
-//   )
-// }
 
 const registerGlobalComponents = (Vue, options) => {
   const globalComponents = Cypress._.get(options, 'extensions.components')
@@ -151,18 +135,18 @@ const resetStoreVM = (Vue, { store }) => {
   return store
 }
 
-function setXMLHttpRequest (w) {
-  // by grabbing the XMLHttpRequest from app's iframe
-  // and putting it here - in the test iframe
-  // we suddenly get spying and stubbing 😁
-  window.XMLHttpRequest = w.XMLHttpRequest
-  return w
-}
+// function setXMLHttpRequest (w) {
+//   // by grabbing the XMLHttpRequest from app's iframe
+//   // and putting it here - in the test iframe
+//   // we suddenly get spying and stubbing 😁
+//   window.XMLHttpRequest = w.XMLHttpRequest
+//   return w
+// }
 
-function setAlert (w) {
-  window.alert = w.alert
-  return w
-}
+// function setAlert (w) {
+//   window.alert = w.alert
+//   return w
+// }
 
 const mountVue = (component, optionsOrProps = {}) => {
   checkMountModeEnabled()
@@ -193,7 +177,6 @@ const mountVue = (component, optionsOrProps = {}) => {
   // const mountId = options.mountId || defaultMountId
   // const el = doc.getElementById(mountId)
 
-
   // set global Vue instance:
   // 1. convenience for debugging in DevTools
   // 2. some libraries might check for this global
@@ -222,43 +205,41 @@ const mountVue = (component, optionsOrProps = {}) => {
   //   copyStyles(component)
   // }
 
-  return cy
-    .window({ log: false })
-    .then((win) => {
-      win.Vue = Vue
+  return cy.window({ log: false }).then(win => {
+    win.Vue = Vue
 
-      const document = cy.state('document')
-      const el = document.getElementById('cypress-jsdom')
-      const componentNode = document.createElement('div')
-      el.append(componentNode)
+    const document = cy.state('document')
+    const el = document.getElementById('cypress-jsdom')
+    const componentNode = document.createElement('div')
+    el.append(componentNode)
 
-      // setup Vue instance
-      installFilters(Vue, options)
-      installMixins(Vue, options)
-      installPlugins(Vue, options)
-      registerGlobalComponents(Vue, options)
-      deleteCachedConstructors(component)
+    // setup Vue instance
+    installFilters(Vue, options)
+    installMixins(Vue, options)
+    installPlugins(Vue, options)
+    registerGlobalComponents(Vue, options)
+    deleteCachedConstructors(component)
 
-      // create root Vue component
-      // and make it accessible via Cypress.vue
-      if (isConstructor(component)) {
-        const Cmp = Vue.extend(component)
-        Cypress.vue = new Cmp(props).$mount(componentNode)
-        // copyStyles(Cmp)
-      } else {
-        Cypress.vue = new Vue(component).$mount(componentNode)
-        // copyStyles(component)
-      }
-    })
-    // .then(setXMLHttpRequest)
-    // .then(setAlert)
-    // .then(() => {
-    //   return cy.wrap(Cypress.vue)
-    // })
+    // create root Vue component
+    // and make it accessible via Cypress.vue
+    if (isConstructor(component)) {
+      const Cmp = Vue.extend(component)
+      Cypress.vue = new Cmp(props).$mount(componentNode)
+      // copyStyles(Cmp)
+    } else {
+      Cypress.vue = new Vue(component).$mount(componentNode)
+      // copyStyles(component)
+    }
+  })
+  // .then(setXMLHttpRequest)
+  // .then(setAlert)
+  // .then(() => {
+  //   return cy.wrap(Cypress.vue)
+  // })
 }
 
 // the double function allows mounting a component quickly
 // beforeEach(mountVue(component, options))
 const mountCallback = (...args) => () => mountVue(...args)
 
-module.exports = {mount: mountVue, mountCallback}
+module.exports = { mount: mountVue, mountCallback }
